+++ scripts/tests/swap-allowance.ts
@@ -1,176 +1,27 @@
 import axios from "axios";
-import { CHAIN_IDs, TOKEN_SYMBOLS_MAP } from "@across-protocol/constants";
-import { ethers, Wallet } from "ethers";
+import { Wallet } from "ethers";
 import dotenv from "dotenv";
 import { getProvider } from "../../api/_utils";
+import {
+  filterTestCases,
+  MIN_OUTPUT_CASES,
+  EXACT_OUTPUT_CASES,
+  SWAP_API_BASE_URL,
+} from "./_swap-utils";
 dotenv.config();
 
 /**
  * Manual test script for the swap API. Should be converted to a proper test suite.
  */
-
-const depositor = "0x9A8f92a830A5cB89a3816e3D267CB7791c16b04D";
-const originChainId = CHAIN_IDs.OPTIMISM;
-const destinationChainId = CHAIN_IDs.ARBITRUM;
-const anyDestinationOutputTokens = {
-  [CHAIN_IDs.ARBITRUM]: "0x74885b4D524d497261259B38900f54e6dbAd2210", // APE
-};
-
-const MIN_OUTPUT_CASES = [
-  // B2B
-  {
-    labels: ["B2B", "MIN_OUTPUT", "USDC - USDC"],
-    params: {
-      amount: ethers.utils.parseUnits("1", 6).toString(),
-      tradeType: "minOutput",
-      inputToken: TOKEN_SYMBOLS_MAP.USDC.addresses[originChainId],
-      originChainId,
-      outputToken: TOKEN_SYMBOLS_MAP.USDC.addresses[destinationChainId],
-      destinationChainId,
-      depositor,
-    },
-  },
-  {
-    labels: ["B2B", "MIN_OUTPUT", "Native ETH - Native ETH"],
-    params: {
-      amount: ethers.utils.parseUnits("0.001", 18).toString(),
-      tradeType: "minOutput",
-      inputToken: ethers.constants.AddressZero,
-      originChainId,
-      outputToken: ethers.constants.AddressZero,
-      destinationChainId,
-      depositor,
-    },
-  },
-  {
-    labels: ["B2B", "MIN_OUTPUT", "USDC - Native ETH"],
-    params: {
-      amount: ethers.utils.parseUnits("0.001", 18).toString(),
-      tradeType: "minOutput",
-      inputToken: TOKEN_SYMBOLS_MAP.USDC.addresses[originChainId],
-      originChainId,
-      outputToken: ethers.constants.AddressZero,
-      destinationChainId,
-      depositor,
-    },
-  },
-  {
-    labels: ["B2B", "MIN_OUTPUT", "Native ETH - USDC"],
-    params: {
-      amount: ethers.utils.parseUnits("3", 6).toString(),
-      tradeType: "minOutput",
-      inputToken: ethers.constants.AddressZero,
-      originChainId,
-      outputToken: TOKEN_SYMBOLS_MAP.USDC.addresses[destinationChainId],
-      destinationChainId,
-      depositor,
-    },
-  },
-  // B2A
-  {
-    labels: ["B2A", "MIN_OUTPUT", "USDC - APE"],
-    params: {
-      amount: ethers.utils.parseUnits("3", 18).toString(),
-      tradeType: "minOutput",
-      inputToken: TOKEN_SYMBOLS_MAP.USDC.addresses[originChainId],
-      originChainId,
-      outputToken: anyDestinationOutputTokens[destinationChainId],
-      destinationChainId,
-      depositor,
-    },
-  },
-  {
-    labels: ["B2A", "MIN_OUTPUT", "USDC - Native ETH"],
-    params: {
-      amount: ethers.utils.parseUnits("0.001", 18).toString(),
-      tradeType: "minOutput",
-      inputToken: TOKEN_SYMBOLS_MAP.USDC.addresses[originChainId],
-      originChainId,
-      outputToken: ethers.constants.AddressZero,
-      destinationChainId,
-      depositor,
-    },
-  },
-  // A2B
-  {
-    labels: ["A2B", "MIN_OUTPUT", "bridged USDC - USDC"],
-    params: {
-      amount: ethers.utils.parseUnits("1", 6).toString(),
-      tradeType: "minOutput",
-      inputToken:
-        TOKEN_SYMBOLS_MAP["USDC.e"].addresses[originChainId] ||
-        TOKEN_SYMBOLS_MAP.USDbC.addresses[originChainId],
-      originChainId,
-      outputToken: TOKEN_SYMBOLS_MAP.USDC.addresses[destinationChainId],
-      destinationChainId,
-      depositor,
-    },
-  },
-  {
-    labels: ["A2B", "MIN_OUTPUT", "USDC - WETH"],
-    params: {
-      amount: ethers.utils.parseUnits("0.001", 18).toString(),
-      tradeType: "minOutput",
-      inputToken: TOKEN_SYMBOLS_MAP.USDC.addresses[originChainId],
-      originChainId,
-      outputToken: TOKEN_SYMBOLS_MAP.WETH.addresses[destinationChainId],
-      destinationChainId,
-      depositor,
-    },
-  },
-  {
-    labels: ["A2B", "MIN_OUTPUT", "WETH - USDC"],
-    params: {
-      amount: ethers.utils.parseUnits("1", 6).toString(),
-      tradeType: "minOutput",
-      inputToken: TOKEN_SYMBOLS_MAP.WETH.addresses[originChainId],
-      originChainId,
-      outputToken: TOKEN_SYMBOLS_MAP.USDC.addresses[destinationChainId],
-      destinationChainId,
-      depositor,
-    },
-  },
-  // A2A
-  {
-    labels: ["A2A", "MIN_OUTPUT", "bridged USDC - APE"],
-    params: {
-      amount: ethers.utils.parseUnits("1", 18).toString(),
-      tradeType: "minOutput",
-      inputToken:
-        TOKEN_SYMBOLS_MAP["USDC.e"].addresses[originChainId] ||
-        TOKEN_SYMBOLS_MAP.USDbC.addresses[originChainId],
-      originChainId,
-      outputToken: anyDestinationOutputTokens[destinationChainId], // APE Coin
-      destinationChainId,
-      depositor,
-    },
-  },
-];
-const EXACT_OUTPUT_CASES = MIN_OUTPUT_CASES.map((testCase) => ({
-  labels: testCase.labels.map((label) => label.replace("MIN", "EXACT")),
-  params: {
-    ...testCase.params,
-    tradeType: "exactOutput",
-  },
-}));
-
 async function swap() {
   const filterString = process.argv[2];
   const testCases = [...MIN_OUTPUT_CASES, ...EXACT_OUTPUT_CASES];
-  const labelsToFilter = filterString ? filterString.split(",") : [];
-  const filteredTestCases = testCases.filter((testCase) => {
-    const matches = labelsToFilter.filter((label) =>
-      testCase.labels
-        .map((label) => label.toLowerCase())
-        .includes(label.toLowerCase())
-    );
-    return matches.length === labelsToFilter.length;
-  });
+  const filteredTestCases = filterTestCases(testCases, filterString);
   for (const testCase of filteredTestCases) {
     console.log("\nTest case:", testCase.labels.join(" "));
     console.log("Params:", testCase.params);
     const response = await axios.get(
-      `http://localhost:3000/api/swap/allowance`,
+      `${SWAP_API_BASE_URL}/api/swap/allowance`,
       {
         params: testCase.params,
       }
